ARI - restructured into mod, bin, table, and test modules
